# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: User Management APIs
host: 'localhost:10004'
basePath: /rbac
schemes:
- http
consumes:
- multipart/form-data
- application/json
produces:
- application/json
- text/plain
paths:
...
  '/usr/reviewpermission/{app}':
    x-swagger-router-controller: controller
    get:
      description: Retrieve if user has review permission for an App
      operationId: ReviewPermission
      parameters:
      - name: authorization
        in: header
        type: string
        description: The JWT token for req. validation
      - name: app
        in: path
        required: true
        type: string
        description: App name against which we db will be querried
      responses:
        '200':
          description: Atleast one service has review permission
        '400':
          description: Bad Parameter
        '500':
          description: Internal server error
  '/usr/reviewpermissionservice/{entity}':
    x-swagger-router-controller: controller
    get:
      description: Retrieve if user has review permission for an Service
      operationId: ReviewPermissionService
      parameters:
      - name: authorization
        in: header
        type: string
        description: The JWT token for req. validation
      - name: entity
        in: path
        required: true
        type: string
        description: service id against which we db will be querried
      - name: user
        in: query
        type: string
        description: user id to check is user has review permission
      responses:
        '200':
          description: Atleast one service has review permission
        '404':
          description: No service with role permission as review
        '500':
          description: Internal server error
...